Refactor to use one actions.ts file, remove unnecessary use-server usage
