+++ src/app/api/files/[file]/route.ts
@@ -1,10 +1,10 @@
 import "server-only";
-import { getFile } from "@/store/files";
-import { getSession, withApiAuthRequired } from "@auth0/nextjs-auth0";
+import { getFileFromStore } from "@/store/files";
+import { getSession } from "@auth0/nextjs-auth0";
 import { readFile } from "fs/promises";
 import mime from "mime";
 import { NextResponse } from "next/server";
-import { canViewFile } from "@/data/authorization";
+import { canViewFile } from "@/app/authorization";
 
 export const dynamic = "force-dynamic";
 export const GET = async function (request, { params }) {
@@ -14,7 +14,7 @@ export const GET = async function (request, { params }) {
 
     // If we're allowed to see the file, return it
     if (await canViewFile(user?.sub, fileId)) {
-      const file = await getFile(params?.file);
+      const file = await getFileFromStore(params?.file);
       const filePath = `${process.cwd()}/upload/${file?.fileName}`;
       const mimeType = mime.getType(filePath);
       const data = await readFile(filePath);
