+++ src/app/authentication.ts
@@ -1,4 +1,4 @@
-"use server";
+import "server-only";
 import { getSession } from "@auth0/nextjs-auth0";
 
 export async function isAuthenticated() {
