+++ src/components/drive/header.tsx
@@ -18,7 +18,7 @@ import {
 import { Label } from "@/components/ui/label";
 import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
-import { createFolder } from "@/actions/folders";
+import { createFolder } from "@/app/actions";
 import { useToast } from "@/components/ui/use-toast";
 import { useState } from "react";
