+++ src/data/user.ts
@@ -1,4 +1,3 @@
-"use server";
 import { getSession } from "@auth0/nextjs-auth0";
 
 export async function getUserId() {
