+++ src/middleware.ts
@@ -1,3 +1,8 @@
 import { withMiddlewareAuthRequired } from "@auth0/nextjs-auth0/edge";
 
 export default withMiddlewareAuthRequired();
+
+export const config = {
+    matcher: '/((?!api/auth).*)'
+};
+
