+++ src/store/files.ts
@@ -1,4 +1,3 @@
-import "server-only";
 import { kv } from "@vercel/kv";
 
 export interface StoredFile {
@@ -9,39 +8,47 @@ export interface StoredFile {
   id?: string;
 }
 
-export async function getFile(id: string) {
+export async function getFileFromStore(id: string) {
   return { ...(await kv.hgetall(`file:${id}`)), id };
 }
 
-export async function getFiles(parent: string) {
+export async function getFilesFromStore(parent: string) {
   const filesForFolder = await kv.smembers(`folder-files:${parent}`);
   const promises: Array<Promise<StoredFile>> = [];
   filesForFolder.forEach((fileId) =>
     promises.push(kv.hgetall(`file:${fileId}`)),
   );
 
-  return (await Promise.all(promises)).map((file, index) => ({
-    ...file,
-    id: filesForFolder[index],
-  }));
+  return (await Promise.all(promises))?.map((file, index) => {
+    if (!!file) {
+      return {
+        ...file,
+        id: filesForFolder[index],
+      }
+    }
+  }).filter(Boolean);;
 }
 
-export async function getFilesSubset(subset: Array<string>) {
+export async function getFilesSubsetFromStore(subset: Array<string>) {
   const promises: Array<Promise<StoredFile>> = [];
   subset.forEach((fileId) => promises.push(kv.hgetall(`file:${fileId}`)));
 
-  return (await Promise.all(promises)).map((file, index) => ({
-    ...file,
-    id: subset[index],
-  }));
+  return (await Promise.all(promises))?.map((file, index) => {
+    if (!!file) {
+      return {
+        ...file,
+        id: subset[index],
+      }
+    }
+  }).filter(Boolean);
 }
 
-export async function createFile(
+export async function createFileInStore(
   fileId: string,
   parent: string,
   file: StoredFile,
 ) {
   await kv.hset(`file:${fileId}`, { ...file });
   await kv.sadd(`folder-files:${parent}`, fileId);
-  return await getFiles(parent);
+  return await getFilesFromStore(parent);
 }
